Keybindings finally work as supposed!
Had to move the keybinding section to the top of the document because otherwise
I'm referencing a keybinding that hasn't been defined yet, and I have to demand
that both General and which-key are loaded from the beginning to make sure they
both work in unison. But damn, it works!